Fail fast if we cannot connect to the kubernetes cluster
Since we currently do not set any timeouts on our kubernetes clients,
it takes CLI a long time before returning an error in case the kubeconfig
is not pointing to a running cluster. This is frustrating for the user.

I added a connection attempt with a 1s timeout before we start our usual
init steps to avoid this.